Confl

+++ PizzaPlaneta/src/main/java/com/pizzaplaneta/PizzaPlaneta/Entity/Bebidas.java
@@ -25,7 +25,7 @@ public class Bebidas {
         return "Prueba";
     }
 
-    public String compra(Long id){
+    public String compra(int id){
         return "Comprado de bebidas";
     }
